Restructuring Codebase and Adding Tests * added queries file and build folder constant * moved appleInfo constants to api directory * moved endpoint callbacks into separate file from server * ported node code to es6 * added jest and supertest to api directory
