+++ api/constants.js
@@ -1,0 +1,57 @@
+export const appleInfo = [
+    {
+        id: 0,
+        name: "HoneyCrisp",
+        picture: "/apple-honeycrisp.png"
+    },
+    {
+        id: 1,
+        name: "Red Delicious",
+        picture: "/apple-red-delicious.png"
+    },
+    {
+        id: 2,
+        name: "Gala",
+        picture: "/apple-gala.png"
+    },
+    {
+        id: 3,
+        name: "Fuji",
+        picture: "/apple-fuji.png"
+    },
+    {
+        id: 4,
+        name: "Granny Smith",
+        picture: "/apple-grannysmith.png"
+    },
+    {
+        id: 5,
+        name: "Golden Delicious",
+        picture: "/apple-golden-delicious.png"
+    },
+    {
+        id: 6,
+        name: "Ambrosia",
+        picture: "/apple-ambrousia.png"
+    },
+    {
+        id: 7,
+        name: "McIntosh",
+        picture: "/apple-mcIntosh.png"
+    },
+    {
+        id: 8,
+        name: "Crispin",
+        picture: "/apple-crispin.png"
+    },
+    {
+        id: 9,
+        name: "JonaGold",
+        picture: "/apple-jonagold.png"
+    },
+    {
+        id: 10,
+        name: "Empire",
+        picture: "/apple-empire.png"
+    }
+]
